+++ src/lib/pages/module-details/components/ModuleInfo.tsx
@@ -5,6 +5,7 @@ import type { MoveVerifyInfoResponse } from "lib/services/types";
 import type {
   IndexedModule,
   ModulePublishInfo,
+  MoveVerifyStatus,
   Nullish,
   Option,
 } from "lib/types";
@@ -15,14 +16,22 @@ export interface ModuleInfoProps {
   indexedModule: IndexedModule;
   modulePublishInfo: Option<ModulePublishInfo>;
   verificationData: Nullish<MoveVerifyInfoResponse>;
+  moveVerifyStatus: MoveVerifyStatus;
 }
 
-export const ModuleInfo = ({ verificationData, ...props }: ModuleInfoProps) => (
+export const ModuleInfo = ({
+  verificationData,
+  moveVerifyStatus,
+  ...props
+}: ModuleInfoProps) => (
   <Flex flexDirection="column" gap={4}>
     <Heading as="h6" variant="h6" fontWeight={600}>
       Module Information
     </Heading>
     <ModuleInfoBody {...props} />
-    <ModuleSourceCode sourceCode={verificationData?.source} />
+    <ModuleSourceCode
+      sourceCode={verificationData?.source}
+      moveVerifyStatus={moveVerifyStatus}
+    />
   </Flex>
 );