+++ src/lib/pages/module-details/components/ModuleTop.tsx
@@ -14,16 +14,16 @@ import { CopyButton } from "lib/components/copy";
 import { CopyLink } from "lib/components/CopyLink";
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import { CustomIcon } from "lib/components/icon";
-import { MoveVerifyStatus } from "lib/components/move-verify-section";
 import { MoveVerifyBadge } from "lib/components/MoveVerifyBadge";
 import { Tooltip } from "lib/components/Tooltip";
 import { UpgradePolicy } from "lib/types";
-import type { Addr, IndexedModule } from "lib/types";
+import type { Addr, IndexedModule, MoveVerifyStatus } from "lib/types";
 import { isHexModuleAddress, isHexWalletAddress, truncate } from "lib/utils";
 
 interface ModuleTopProps {
   moduleData: IndexedModule;
   isVerified: boolean;
+  moveVerifyStatus: MoveVerifyStatus;
 }
 
 const baseTextStyle: TextProps = {
@@ -154,7 +154,11 @@ const ModuleCta = ({
   );
 };
 
-export const ModuleTop = ({ moduleData, isVerified }: ModuleTopProps) => {
+export const ModuleTop = ({
+  moduleData,
+  isVerified,
+  moveVerifyStatus,
+}: ModuleTopProps) => {
   const isMobile = useMobile();
 
   const { convertHexWalletAddress, convertHexModuleAddress } =
@@ -212,7 +216,7 @@ export const ModuleTop = ({ moduleData, isVerified }: ModuleTopProps) => {
           </Heading>
           {/* TODO: Add status & remove isVerified */}
           <Tooltip label="This module's verification is supported by its provided source code.">
-            <MoveVerifyBadge status={MoveVerifyStatus.Verified} />
+            <MoveVerifyBadge status={moveVerifyStatus} />
             {isVerified}
           </Tooltip>
         </Flex>