+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- [#1101](https://github.com/alleslabs/celatone-frontend/pull/1101) Add module verification badge and entry points
 - [#1093](https://github.com/alleslabs/celatone-frontend/pull/1093) My past module verifications with functionalities
 - [#1085](https://github.com/alleslabs/celatone-frontend/pull/1085) Add my module verification details
 - [#1087](https://github.com/alleslabs/celatone-frontend/pull/1087) Add modules verify page